simplify object-storage
