+++ packages/infrastructure/src/index.ts
@@ -1,1 +1,1 @@
-export { getObject } from '@/object-storage/queries/get-object';
+export { getObject } from '@/object-storage';
